+++ example_test.go
@@ -8,8 +8,8 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/goyek/goyek/v2"
-	"github.com/goyek/goyek/v2/middleware"
+	"github.com/goyek/goyek/v3"
+	"github.com/goyek/goyek/v3/middleware"
 )
 
 var (