+++ middleware/reportlongrun_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/goyek/goyek/v2"
-	"github.com/goyek/goyek/v2/middleware"
+	"github.com/goyek/goyek/v3"
+	"github.com/goyek/goyek/v3/middleware"
 )
 
 func TestReportLongRun(t *testing.T) {