+++ status_test.go
@@ -3,7 +3,7 @@ package goyek_test
 import (
 	"testing"
 
-	"github.com/goyek/goyek/v2"
+	"github.com/goyek/goyek/v3"
 )
 
 func TestStatus_String(t *testing.T) {