+++ CHANGELOG.md
@@ -8,8 +8,9 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 
 ## [Unreleased](https://github.com/goyek/goyek/compare/v2.2.0...HEAD)
 
-- Added `A.WithContext` to support changing the context of an action by the user.
-  The intended use is to pass task parameters via context.
+- Add `A.WithContext` that creates a derived `A` with a changed context.
+  Thanks to it `A` can be reused to pass cancelation and values via context
+  to the helper functions.
 
 ### Removed