+++ CHANGELOG.md
@@ -13,6 +13,8 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 - Add `A.WithContext` that creates a derived `A` with a changed context.
   Thanks to it `A` can be reused to pass cancelation and values via context
   to the helper functions.
+- Add `A.Chdir` that changes the current working directory and during action cleanup
+  restores it to its original value.
 
 ### Removed